"Fix" issue that "monitor set" needed to be called twice
because the first time VG_(strtod) returns a strange double
value, by setting the optimization level of VG_(strtod) to O0.
No idea what the actual problem is.
That's how I found the "fix": I ran valgrind in the debugger and
the local variables frac and x changed in a strange way in the
second while(is_dec_digit(...)) loop. Trying to set a memory
watchpoint, I found that I cannot dereference them in gdb.
VG_(printf)'ing the addresses in the code sometimes (?) fixed
the issue.