WrTmp/Put/Store/RdTmp/... gradient for Ity_I32 also
Apparantly, when we copy double values in the C source, the
VEX IR code sometimes makes two copies of Ity_I32 variables.
Therefore we now assign 8-byte derivative information to them
as well:
The derivative of an Ity_I32 "half" of a Ity_F64 value must be
the derivative of the Ity_F64 value, and for all other Ity_I32
values the derivative can be anything. We must be careful in
differentiate_expr to not take such "anything" derivatives into
account.
When Store'ing an Ity_I32 value, the derivative is stored in the
shadow memory aligned to 8 byte.
When Put'ing an Ity_I32 value, the same rule applies. I don't know
whether this is fully legitimate. PutI'ing an Ity_I32 value is
still a TODO.