Replace storeShadowMemory in dg_dot_store, dg_bar_store,
and likewise loadShadowMemory in dg_dot_load, dg_bar_load.
Previously, we had to partition and convert the data into I64
portions that we can pass as arguments to dirty calls. Now, we
use some portion of memory allocated during the initialization of
shadow memory. A dedicated register would maybe be faster, but
we are already using the full guest state with regular shadow
data.